Extract "handlers".
